Debugging to get sim to run and a torque change (Kyle) Parts of this commit needs to be looked at as I do not know if the impleimentation of the external reduction in `control_node.cpp` is correct

+++ flapping_proj/optimization/geometry_classes.py
@@ -139,7 +139,7 @@ class LineSegments(Curve):
             return num_segments * del_y
 
         def z(t: float):
-            index = int((np.floor(t * num_segments) - np.floor(t)))
+            index = (np.floor(t * num_segments) - np.floor(t)).astype(int)
             del_z = z_points[index + 1] - z_points[index]
             return del_z * (num_segments * t - index) + z_points[index]
